requirements.txt
openai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.64.0
requests>=2.27.0